win32ole.c: suppress warning
* ext/win32ole/win32ole.c (ole_rec2variant): cast to suppress a
  format warning on x86_64-cygwin.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@46981 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/win32ole/win32ole.c
@@ -1710,7 +1710,7 @@ ole_rec2variant(VALUE rec, VARIANT *var)
         }
         prec->pdata = ALLOC_N(char, size);
         if (!prec->pdata) {
-            rb_raise(rb_eRuntimeError, "failed to memory allocation of %ld bytes", size);
+            rb_raise(rb_eRuntimeError, "failed to memory allocation of %lu bytes", (unsigned long)size);
         }
         hr = pri->lpVtbl->RecordInit(pri, prec->pdata);
         if (FAILED(hr)) {
